Changed MessageID to be a long instead of int to match signal specifications
git-svn-id: file:///e/SVN/serious_messenger@13 b1debafe-a605-4e1b-80f9-83a58a4e6cd1

+++ Serious Client/src/client/Conversation.java
@@ -86,13 +86,13 @@ public class Conversation {
 	}
 	
 	// TODO: Finish me
-	public boolean receiveMessage(String userName, int id, String message) {
+	public boolean receiveMessage(String userName, long id, String message) {
 		if(userName == null || message == null) { return false; }
 		return receiveMessage(getParticipant(userName), id, message);
 	}
 	
 	// TODO: Finish me
-	public boolean receiveMessage(Contact c, int id, String message) {
+	public boolean receiveMessage(Contact c, long id, String message) {
 		if(c == null || message == null) { return false; }
 		
 		Message m = new Message(c, id, message);
